提交

+++ src/main/java/com/bsq/fresh/Vo/entityVO/CategoryTwoFoodVO.java
@@ -3,6 +3,7 @@ package com.bsq.fresh.Vo.entityVO;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -18,7 +19,7 @@ public class CategoryTwoFoodVO {
     private String categoryTwoFoodDesc;
 
     @JsonProperty("marketPrice")
-    private double categoryTwoFoodPrice;
+    private BigDecimal categoryTwoFoodPrice;
 
     @JsonProperty("originPrice")
     private double categoryTwoFoodOriginPrice;
